Added friends pane
