+++ src/main/resources/com/toysocialnetworkgui/controller/friends.fxml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+
+<AnchorPane prefHeight="650.0" prefWidth="800.0" stylesheets="@../../../style/style.css" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.toysocialnetworkgui.controller.FriendsController">
+   <children>
+      <VBox layoutX="121.0" layoutY="149.0" prefHeight="351.0" prefWidth="558.0" spacing="10.0">
+         <children>
+            <HBox spacing="5.0">
+               <children>
+                  <TextField fx:id="textFieldSearchFriend" prefHeight="26.0" prefWidth="168.0" promptText="search friend" />
+                  <HBox spacing="197.0">
+                     <children>
+                        <Button fx:id="buttonSearch" mnemonicParsing="false" onAction="#onButtonSearchFriend" styleClass="image_wrapper">
+                           <graphic>
+                              <ImageView fitHeight="30.0" fitWidth="30.0" pickOnBounds="true" preserveRatio="true">
+                                 <image>
+                                    <Image url="@../../../images/magnifier.png" />
+                                 </image>
+                              </ImageView>
+                           </graphic>
+                        </Button>
+                        <ComboBox fx:id="comboBoxMonth" onAction="#onSelectMonth" prefHeight="26.0" prefWidth="159.0" promptText="choose month" />
+                     </children>
+                  </HBox>
+               </children>
+            </HBox>
+            <TableView fx:id="tableViewFriends" onMouseClicked="#handleTableClickEvent" prefHeight="272.0" prefWidth="558.0">
+              <columns>
+                  <TableColumn fx:id="tableColumnEmail" prefWidth="75.0" text="email" visible="false" />
+                <TableColumn fx:id="tableColumnFirstname" maxWidth="155.0" minWidth="155.0" prefWidth="155.0" resizable="false" text="First name" />
+                <TableColumn fx:id="tableColumnLastname" maxWidth="155.0" minWidth="155.0" prefWidth="155.0" resizable="false" text="Last name" />
+                  <TableColumn fx:id="tableColumnDate" maxWidth="155.0" minWidth="155.0" prefWidth="155.0" resizable="false" text="Friends from" />
+                  <TableColumn fx:id="tableColumnRemove" maxWidth="45.0" minWidth="45.0" prefWidth="45.0" resizable="false" />
+                  <TableColumn fx:id="tableColumnMessage" maxWidth="45.0" minWidth="45.0" prefWidth="45.0" resizable="false" />
+              </columns>
+            </TableView>
+            <HBox spacing="148.0">
+               <children>
+                  <Button fx:id="buttonPreviousPage" mnemonicParsing="false" onAction="#onPreviousPageButtonClick" styleClass="image_wrapper">
+                     <graphic>
+                        <ImageView fitHeight="30.0" fitWidth="30.0" pickOnBounds="true" preserveRatio="true">
+                           <image>
+                              <Image url="@../../../images/prev_page.png" />
+                           </image>
+                        </ImageView>
+                     </graphic>
+                  </Button>
+                  <Label fx:id="labelPage" alignment="CENTER" contentDisplay="CENTER" minWidth="-Infinity" prefHeight="18.0" prefWidth="200.0" text="page" />
+                  <Button fx:id="buttonNextPage" mnemonicParsing="false" onAction="#onNextPageButtonClick" styleClass="image_wrapper">
+                     <graphic>
+                        <ImageView fitHeight="30.0" fitWidth="30.0" pickOnBounds="true" preserveRatio="true">
+                           <image>
+                              <Image url="@../../../images/next_page.png" />
+                           </image>
+                        </ImageView>
+                     </graphic>
+                  </Button>
+               </children>
+            </HBox>
+         </children>
+      </VBox>
+   </children>
+</AnchorPane>
