+++ src/main/resources/com/toysocialnetworkgui/controller/loggedScene.fxml
@@ -3,11 +3,7 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.Cursor?>
 <?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.SplitPane?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.control.TextField?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
@@ -87,7 +83,7 @@
                               </ImageView>
                            </graphic>
                         </Button>
-                        <Button alignment="CENTER_LEFT" minHeight="60.0" minWidth="120.0" mnemonicParsing="false" prefHeight="22.0" prefWidth="119.0" styleClass="category" text="Friends" textAlignment="CENTER">
+                        <Button fx:id="buttonFriends" alignment="CENTER_LEFT" minHeight="60.0" minWidth="120.0" mnemonicParsing="false" onAction="#onButtonFriendsClick" prefHeight="22.0" prefWidth="119.0" styleClass="category" text="Friends" textAlignment="CENTER">
                            <graphic>
                               <ImageView fitHeight="30.0" fitWidth="30.0" pickOnBounds="true" preserveRatio="true" styleClass="imageWrapper">
                                  <image>
@@ -103,11 +99,6 @@
                            </font>
                         </Button>
                         <Button fx:id="buttonAddFriend" mnemonicParsing="false" onAction="#onAddFriendButtonClick" prefHeight="26.0" prefWidth="98.0" text="Add friend">
-                           <VBox.margin>
-                              <Insets bottom="5.0" top="5.0" />
-                           </VBox.margin>
-                        </Button>
-                        <Button fx:id="buttonRemoveFriend" mnemonicParsing="false" onAction="#onRemoveFriendButtonClick" prefHeight="26.0" prefWidth="100.0" text="Remove friend">
                            <VBox.margin>
                               <Insets bottom="5.0" top="5.0" />
                            </VBox.margin>
@@ -226,43 +217,6 @@
                         <Image url="@../../../images/meme.jpg" />
                      </image>
                   </ImageView>
-                  <VBox alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" visible="false">
-                     <children>
-                        <HBox>
-                           <children>
-                              <ComboBox fx:id="comboBoxMonth" onAction="#onSelectMonth" prefHeight="26.0" prefWidth="163.0" promptText="filter friends by month" visibleRowCount="13" />
-                              <HBox prefHeight="100.0" prefWidth="200.0">
-                                 <children>
-                                    <TextField fx:id="textFieldSearchFriend" onAction="#clearSearchFriendSelection" promptText="Search">
-                                       <HBox.margin>
-                                          <Insets left="25.0" right="5.0" />
-                                       </HBox.margin>
-                                    </TextField>
-                                    <Button fx:id="buttonSearch" mnemonicParsing="false" onAction="#onButtonSearchFriend" prefHeight="26.0" prefWidth="30.0" styleClass="buttonSearch" />
-                                 </children>
-                              </HBox>
-                           </children>
-                        </HBox>
-                        <TableView fx:id="tableViewFriends" maxHeight="-Infinity" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="521.0">
-                          <columns>
-                              <TableColumn fx:id="tableColumnEmail" prefWidth="75.0" text="email" visible="false" />
-                            <TableColumn fx:id="tableColumnFirstname" prefWidth="86.40002286434174" sortable="false" text="First name" />
-                              <TableColumn fx:id="tableColumnLastname" prefWidth="75.0" sortable="false" text="Last name" />
-                            <TableColumn fx:id="tableColumnDate" prefWidth="72.0" sortType="DESCENDING" text="Friends from" />
-                          </columns>
-                           <columnResizePolicy>
-                              <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-                           </columnResizePolicy>
-                        </TableView>
-                        <HBox maxWidth="1.7976931348623157E308" prefHeight="100.0" prefWidth="200.0">
-                           <children>
-                              <Button fx:id="previousPage" maxWidth="1.7976931348623157E308" minWidth="-Infinity" mnemonicParsing="false" onAction="#onPreviousPageButtonClick" prefHeight="26.0" prefWidth="138.0" styleClass="buttonPage" text="Previous page" />
-                              <VBox prefHeight="100.0" prefWidth="383.0" />
-                              <Button fx:id="nextPage" maxWidth="1.7976931348623157E308" minWidth="-Infinity" mnemonicParsing="false" onAction="#onNextPageButtonClick" prefHeight="26.0" prefWidth="100.0" styleClass="buttonPage" text="Next page" textAlignment="CENTER" />
-                           </children>
-                        </HBox>
-                     </children>
-                  </VBox>
                </children>
             </StackPane>
          </children>
