+++ src/main/resources/style/style.css
@@ -26,6 +26,21 @@
     -fx-text-fill: white;
 
 }
+
+.image_wrapper{
+    -fx-color: -fx-base ;
+    -fx-background-color: transparent;
+    -fx-background-insets: 0 ;
+    -fx-background-radius: 0 ;
+    -fx-padding: 0 ;
+    -fx-text-fill: #a8a5a5 ;
+    -fx-alignment: center;
+    -fx-min-width: 30px;
+    -fx-max-width: 30px;
+    -fx-min-height: 30px;
+    -fx-max-height: 30px;
+}
+
 #button:hover{
    -fx-effect: dropshadow( gaussian , rgba( 0, 0, 0, 0.5 ), 10, 0.1, 3, 3);
